requirements.txt
colorlog==3.1.4
confusable-homoglyphs==3.2.0
Django==2.1.2
#-e git+https://github.com/yourlabs/django-autocomplete-light.git@24bff6bafe148ab4ce823ee077f3e62bf80472e3#egg=django_autocomplete_light
django-autocomplete-light
django-bootstrap4==0.0.7
django-braces==1.13.0
django-cors-middleware==1.3.1
django-debug-toolbar==1.10.1
django-redis==4.9.0
django-redis-cache==1.7.1
django-registration==3.0
django-suit==0.2.26
django-waffle==0.14.0
gevent==1.3.7
greenlet==0.4.15
grequests==0.3.0
huey==1.10.3
idna==2.7
oauth2-provider==0.0
oauthlib==2.1.0
gunicorn==19.9.0
peewee==3.7.1
pycparser==2.19
PyMySQL==0.9.2
python-dateutil==2.7.3
pytz==2018.5
raven==6.9.0
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
shortuuid==0.5.0
simpledb==0.4.2
simplejson==3.16.0
six==1.11.0
urllib3==1.23
sentry-sdk==0.5.5